Migrate to LALRPOP for parsing.
Previously, I hand-rolled a scanner and parser for this compiler.
While it worked well enough and _might_ be more performant, the
ease of adding new parts to the grammar became excessively complex,
so I've opted to instead use LALRPOP, a parser generator, to
generate the parser for us. In the long term, I think that will
allow for more time to be spent on parts of the language that
frankly are more important. I doubt parsing will realistically
be a bottleneck, so I've opted _not_ to use a home-rolled parser
at this time. It also has _significantly_ clarified the compilation
steps, which will make it easier to add additional passes in the
future (such as for type checking).