Refactor to use static compilation with Cranelift.
From pretty early on, I realized that the performance of my
hand-rolled VM was not going to cut it and didn't really
match what I wanted out of this project. Instead, I've opted
to bite the bullet and make this whole thing statically compiled.

As you can tell, that's a pretty fundamental shift in the way
this project was originally set up, but for a good reason.
Static compilation will allow for better performance optimization,
portability to different backends (looking at you, WASM), and
avoids the overhead of requiring a VM. While that means that I've
had to take time way from properly finishing the language, it'll
be easier to set this up while things are still early on.